Safari-611.1.10/LayoutTests/http/tests/navigation/anchor-blank-target-implies-rel-noopener-expected.txt

CONSOLE MESSAGE: PASS: New window should not have an opener
Tests that a new window opened via target=_blank does not have an opener
